Styled checkboxes, hashed out most of the js, added buttons to clear and submit maps.
Changed the format of the checkboxes so that they could be styled. Added all the basic javascript in; the maplist functionally works by taking the maps selected in the html and generating a random list of them, alternating in a generated random mode order (ie one SZ, one TC, ect). Next, plan to add functionality to the other buttons which will clear all selected maps, only show the "Splat Zones" column, and choose a set of predetermined maps to make a list out of.